backend/pyproject.toml
[tool.poetry]
name = "digital-prefect-backend"
version = "1.0.0"
description = "Digital journal API"
authors = ["penkhasoveg <pen.egor2002@gmail.com>"]


[tool.black]
line-length = 79


[tool.isort]
multi_line_output=3
include_trailing_comma=true
use_parentheses=true
line_length=79


[tool.taskipy.tasks]
manage = "python manage.py"
clear = "task manage clear"
dumpdata = "task manage dump"
migrate = "task manage migrate --noinput"
server = "task manage runserver"
loaddata = "task manage loaddata"
startapp = "task manage createapp"
initconfig = "task manage initconfig"
makemigrations = "task manage makemigrations"
createsuperuser = "task manage createsuperuser"
collectstatic = "task manage collectstatic --noinput"
gunicorn = "gunicorn --timeout 300 --chdir config.wsgi:application -b 0.0.0.0:8000"
defaultadmin = "task manage defaultadmin"
defaultfixtures = "task manage defaultfixtures"


[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.5"
django-cleanup = "^5.1.0"
django-widget-tweaks = "^1.4.5"
django-settings-export = "^1.2.1"
djangorestframework = "^3.11.1"
django-cors-headers = "^3.6.0"
drf-yasg = "^1.17.0"
arrow = "^0.17.0"
django-import-export = "^2.0.1"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"
mysqlclient = "^2.0.1"
psycopg2 = "^2.8.4"
pillow = "^8.1.0"
django-storages = "^1.11.1"
django-etc = "^1.3.0"
django-template-block-args = "^0.0.5"
djangorestframework-camel-case = "^1.1.2"
django-model-utils = "^4.0.0"
django-filter = "^2.2.0"
django-weasyprint = "^1.0.1"
docxtpl = "^0.11.0"
num2words = "^0.5.10"
xlsxwriter = "^1.2.9"
django-easy-select2 = "^1.5.7"
py-linq = "^1.2.4"
tqdm = "^4.48.2"
django-service-objects = "^0.7.0"
awesome-slugify = "^1.6.5"
django-admin-interface = "^0.17.3"
django-apptemplates = "^1.5"
django-extra-settings = "^0.1.3"


[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
django-debug-toolbar = "^3.1.1"
autopep8 = "^1.5"
django-extensions = "^3.0.9"
pylint-django = "^2.0.13"
taskipy = "^1.1.3"
flake8 = "^3.7.9"
black = "^20.8b1"
Werkzeug = "^1.0.1"
safety = "^1.9.0"
bandit = "^1.7.0"
sphinx = "^3.3.1"
sphinx-rtd-theme = "^0.5.0"
